[refactor] narrow responsibility of fragment
