+++ feature-addhighlight/build.gradle
@@ -53,14 +53,8 @@ dependencies {
     implementation "androidx.core:core-ktx:$android_ktx_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "com.google.android.material:material:$material_version"
-    implementation "androidx.constraintlayout:constraintlayout:$constraintlayout_version"
     implementation "androidx.navigation:navigation-fragment-ktx:$fragment_nav_version"
     implementation "com.jakewharton.timber:timber:$timber_version"
-
-    implementation "androidx.camera:camera-core:${camerax_version}"
-    implementation "androidx.camera:camera-lifecycle:${camerax_version}"
-    implementation "androidx.camera:camera-view:${camerax_version}"
-    implementation "androidx.camera:camera-extensions:${camerax_version}"
 
     implementation "com.google.android.gms:play-services-mlkit-text-recognition:${mlkit_version}"
 
@@ -69,6 +63,7 @@ dependencies {
     implementation 'androidx.compose.material3:material3'
     implementation 'androidx.compose.ui:ui-tooling-preview'
     implementation "androidx.lifecycle:lifecycle-runtime-compose:$lifecycle_version"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:$lifecycle_version"
     implementation "androidx.activity:activity-compose:$activity_version"
     debugImplementation 'androidx.compose.ui:ui-tooling'
     androidTestImplementation composeBom
